requirements.txt
numpy
sympy
pyeda
matplotlib
pyyaml
git+git://github.com/yuce/pyswip@9642fd9#egg=pyswip
pybullet
networkx
pydot
trimesh
minio